+++ module/Core/config/event_dispatcher.config.php
@@ -53,7 +53,7 @@ return [
             'em',
             'Logger_Shlink',
             'config.url_shortener.visits_webhooks',
-            'config.url_shortener.domain',
+            ShortUrl\Transformer\ShortUrlDataTransformer::class,
             Options\AppOptions::class,
         ],
         EventDispatcher\NotifyVisitToMercure::class => [